[Improve] performance

+++ evaluation/evaluation.h
@@ -46,9 +46,9 @@ private:
 	}
 	int __evaluate(chessboard &, int turn, const int row, const int col, bool pure) noexcept;
 	void evaluate_point(chessboard &board, int row, int col) noexcept;
-	const int STWO = 1, STHREE = 2, SFOUR = 3, TWO = 4, THREE = 5, FOUR = 6, FIVE = 7, DFOUR = 8,
+	const uint8_t STWO = 1, STHREE = 2, SFOUR = 3, TWO = 4, THREE = 5, FOUR = 6, FIVE = 7, DFOUR = 8,
 		FOURT = 9, DTHREE = 10, NOTYPE = 11, ANALYSED = 255, TODO = 0;
-	const int BLACK = 1, WHITE = 2;
-	const std::array<int, 3> nturn = { 0, 2, 1 };
-	std::array<int, 15> result;
+	const uint8_t BLACK = 1, WHITE = 2;
+	const std::array<uint8_t, 3> nturn = { 0, 2, 1 };
+	std::array<uint8_t, 15> result;
 };
